feat: cargo install astrid installs both CLI and daemon
Add astrid-daemon as a second [[bin]] entry point in the astrid (CLI)
crate. `cargo install astrid` now installs both binaries. The standalone
astrid-daemon crate stays for workspace development.

Also adds tokio signal feature to workspace deps (needed by daemon).

Closes #612
